move reducer into own file
